Adjusted runtime path to new folder structure.
