+++ Composer/packages/electron-server/package.json
@@ -12,9 +12,9 @@
     "build": "tsc -p tsconfig.build.json && ncp src/preload.js build/preload.js",
     "clean": "rimraf build && rimraf dist",
     "copy-plugins": "node scripts/copy-plugins.js",
-    "copy-templates": "node scripts/copy-templates.js",
+    "copy-runtime": "node scripts/copy-runtime.js",
     "dist": "node scripts/electronBuilderDist.js",
-    "dist:full": "yarn clean && yarn build && yarn run pack && yarn copy-templates && yarn copy-plugins && yarn dist",
+    "dist:full": "yarn clean && yarn build && yarn run pack && yarn copy-runtime && yarn copy-plugins && yarn dist",
     "lint": "eslint --quiet --ext .js,.jsx,.ts,.tsx ./src",
     "lint:fix": "yarn lint --fix",
     "pack": "node scripts/electronBuilderPack.js",
