+++ Composer/packages/electron-server/src/main.ts
@@ -101,7 +101,7 @@ async function loadServer() {
   if (!isDevelopment) {
     // only change paths if packaged electron app
     const unpackedDir = getUnpackedAsarPath();
-    process.env.COMPOSER_RUNTIME_FOLDER = join(unpackedDir, 'BotProject', 'Templates');
+    process.env.COMPOSER_RUNTIME_FOLDER = join(unpackedDir, 'runtime');
     pluginsDir = join(unpackedDir, 'build', 'plugins');
   }
